requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
gpiozero==2.0.1
Pillow==10.1.0
